--- dev/com.ibm.ws.transaction.cloud_fat.dualDB/.classpath
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" path="fat/src" including="com/ibm/ws/transaction/test/FATSuite.java"/>
-	<classpathentry kind="src" path="com.ibm.ws.transaction.cloud_fat.base-fatsrc" excluding="com/ibm/ws/transaction/test/FATSuite.java"/>
-	<classpathentry kind="src" path="com.ibm.ws.transaction.cloud_fat.base-transaction"/>
-	<classpathentry kind="con" path="aQute.bnd.classpath.container"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8">
-		<attributes>
-			<attribute name="module" value="true"/>
-		</attributes>
-	</classpathentry>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
