--- dev/com.ibm.ws.transaction.cloud_fat.dualDB/.gitignore
@@ -1,3 +1,0 @@
-/publish
-/test-applications
-/test-bundles
