--- dev/com.ibm.ws.transaction.cloud_fat.dualDB/.settings/bndtools.core.prefs
@@ -1,2 +1,0 @@
-compileErrorAction=build
-eclipse.preferences.version=1
