--- dev/com.ibm.ws.transaction.cloud_fat.dualDB/.settings/org.eclipse.core.resources.prefs
@@ -1,4 +1,0 @@
-#Ant properties
-#Automatically generated by the ant prepare.settings.files task
-eclipse.preferences.version=1
-encoding/<project>=UTF-8
