+++ dev/com.ibm.ws.transaction.cloud_fat.postgresql/fat/src/com/ibm/ws/transaction/test/.gitignore
@@ -1,1 +1,2 @@
 /tests
+/dbrotationtests
