--- dev/com.ibm.ws.transaction.cloud_fat.simpleDB/.gitignore
@@ -1,2 +1,0 @@
-/test-applications
-/test-bundles
