--- dev/com.ibm.ws.transaction.cloud_fat.simpleDB/build.gradle
@@ -1,56 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2020, 2021 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *     IBM Corporation - initial API and implementation
- *******************************************************************************/
-
-// Define G:A:V coordinates of each dependency
-dependencies {
-  requiredLibs project(":com.ibm.ws.tx.embeddable"),
-               project(":com.ibm.ws.transaction.cloud_fat.base"),
-               'commons-logging:commons-logging:1.1.3'
-}
-
-
-File sharedDir = new File(autoFvtDir, 'publish/shared/resources')
-
-task addDerbyToSharedDir(type: Copy) {
-  from configurations.derby
-  into new File(sharedDir, 'derby')
-  rename 'derby-.*.jar', 'derby.jar'
-}
-
-task copyCommonFiles {
-  doLast {
-    /*
-     * Copy test application resources.
-     */
-    copy {
-      from project(':com.ibm.ws.transaction.cloud_fat.base').file('test-applications')
-      into new File('test-applications')
-    }
-  }
-}
-
-addRequiredLibraries.dependsOn addDerbyToSharedDir
-addRequiredLibraries.dependsOn addJakartaTransformer
-
-jar {
-  dependsOn copyCommonFiles
-}
-
-autoFVT.dependsOn copyCommonFiles
-
-clean.doLast {
-  if (file('test-applications').exists()){
-    file('test-applications').deleteDir()
-  }
-  if (file('build').exists()){
-    file('build').deleteDir()
-  }
-}
