--- dev/com.ibm.ws.transaction.cloud_fat.simpleDB/fat/src/com/ibm/ws/transaction/test/tests/Simple2PCCloudTest.java
@@ -1,277 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2019, 2021 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *     IBM Corporation - initial API and implementation
- *******************************************************************************/
-package com.ibm.ws.transaction.test.tests;
-
-import org.junit.After;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import com.ibm.tx.jta.ut.util.LastingXAResourceImpl;
-import com.ibm.websphere.simplicity.ProgramOutput;
-import com.ibm.websphere.simplicity.ShrinkHelper;
-import com.ibm.websphere.simplicity.log.Log;
-import com.ibm.ws.transaction.web.Simple2PCCloudServlet;
-
-import componenttest.annotation.AllowedFFDC;
-import componenttest.annotation.ExpectedFFDC;
-import componenttest.annotation.Server;
-import componenttest.annotation.TestServlet;
-import componenttest.custom.junit.runner.FATRunner;
-import componenttest.topology.impl.LibertyServer;
-import componenttest.topology.utils.FATServletClient;
-
-@RunWith(FATRunner.class)
-public class Simple2PCCloudTest extends FATServletClient {
-
-    public static final String APP_NAME = "transaction";
-    public static final String SERVLET_NAME = APP_NAME + "/Simple2PCCloudServlet";
-    protected static final int cloud2ServerPort = 9992;
-    private static final long LOG_SEARCH_TIMEOUT = 300000;
-
-    @Server("CLOUD001")
-    @TestServlet(servlet = Simple2PCCloudServlet.class, contextRoot = APP_NAME)
-    public static LibertyServer server1;
-
-    @Server("CLOUD002")
-    @TestServlet(servlet = Simple2PCCloudServlet.class, contextRoot = APP_NAME)
-    public static LibertyServer server2;
-
-    @Server("longLeaseLengthServer1")
-    @TestServlet(servlet = Simple2PCCloudServlet.class, contextRoot = APP_NAME)
-    public static LibertyServer longLeaseLengthServer1;
-
-    @BeforeClass
-    public static void setUp() throws Exception {
-
-        // Create a WebArchive that will have the file name 'app1.war' once it's written to a file
-        // Include the 'app1.web' package and all of it's java classes and sub-packages
-        // Automatically includes resources under 'test-applications/APP_NAME/resources/' folder
-        // Exports the resulting application to the ${server.config.dir}/apps/ directory
-        ShrinkHelper.defaultApp(server1, APP_NAME, "com.ibm.ws.transaction.*");
-        ShrinkHelper.defaultApp(server2, APP_NAME, "com.ibm.ws.transaction.*");
-        ShrinkHelper.defaultApp(longLeaseLengthServer1, APP_NAME, "com.ibm.ws.transaction.*");
-
-        server1.setServerStartTimeout(LOG_SEARCH_TIMEOUT);
-        server2.setServerStartTimeout(LOG_SEARCH_TIMEOUT);
-        server2.setHttpDefaultPort(cloud2ServerPort);
-        longLeaseLengthServer1.setServerStartTimeout(LOG_SEARCH_TIMEOUT);
-    }
-
-    @After
-    public void cleanup() throws Exception {
-
-        // Clean up XA resource files
-        server1.deleteFileFromLibertyInstallRoot("/usr/shared/" + LastingXAResourceImpl.STATE_FILE_ROOT);
-
-        // Remove tranlog DB
-        server1.deleteDirectoryFromLibertyInstallRoot("/usr/shared/resources/data");
-    }
-
-    /**
-     * Test access to the Lease table.
-     *
-     * This is a readiness check to verify that resources are available and accessible.
-     *
-     * @throws Exception
-     */
-    @Test
-    public void testLeaseTableAccess() throws Exception {
-        final String method = "testLeaseTableAccess";
-        StringBuilder sb = null;
-        String id = "001";
-
-        // Start Server1
-        server1.startServer();
-
-        try {
-            sb = runTestWithResponse(server1, SERVLET_NAME, "testLeaseTableAccess");
-
-        } catch (Throwable e) {
-        }
-        Log.info(this.getClass(), method, "testLeaseTableAccess" + id + " returned: " + sb);
-
-        // "CWWKE0701E" error message is allowed
-        server1.stopServer("CWWKE0701E");
-    }
-
-    /**
-     * The purpose of this test is as a control to verify that single server recovery is working.
-     *
-     * The Cloud001 server is started and halted by a servlet that leaves an indoubt transaction.
-     * Cloud001 is restarted and transaction recovery verified.
-     *
-     * @throws Exception
-     */
-    @Test
-    @AllowedFFDC(value = { "javax.transaction.xa.XAException", "com.ibm.ws.recoverylog.spi.RecoveryFailedException" })
-    public void testDBBaseRecovery() throws Exception {
-        final String method = "testDBBaseRecovery";
-        StringBuilder sb = null;
-        String id = "001";
-        // Start Server1
-        server1.startServer();
-
-        try {
-            // We expect this to fail since it is gonna crash the server
-            sb = runTestWithResponse(server1, SERVLET_NAME, "setupRec" + id);
-        } catch (Throwable e) {
-        }
-        Log.info(this.getClass(), method, "setupRec" + id + " returned: " + sb);
-
-        server1.waitForStringInLog("Dump State:");
-
-        // Now re-start cloud1
-        ProgramOutput po = server1.startServerAndValidate(false, true, true);
-        if (po.getReturnCode() != 0) {
-            Log.info(this.getClass(), method, po.getCommand() + " returned " + po.getReturnCode());
-            Log.info(this.getClass(), method, "Stdout: " + po.getStdout());
-            Log.info(this.getClass(), method, "Stderr: " + po.getStderr());
-            Exception ex = new Exception("Could not restart the server");
-            Log.error(this.getClass(), "recoveryTest", ex);
-            throw ex;
-        }
-
-        // Server appears to have started ok. Check for key string to see whether recovery has succeeded
-        server1.waitForStringInTrace("Performed recovery for cloud001");
-
-        // Lastly stop server1
-        // "WTRN0075W", "WTRN0076W", "CWWKE0701E" error messages are expected/allowed
-        server1.stopServer("WTRN0075W", "WTRN0076W", "CWWKE0701E");
-
-        // Lastly, clean up XA resource file
-        server1.deleteFileFromLibertyInstallRoot("/usr/shared/" + LastingXAResourceImpl.STATE_FILE_ROOT);
-    }
-
-    /**
-     * The purpose of this test is to verify simple peer transaction recovery.
-     *
-     * The Cloud001 server is started and halted by a servlet that leaves an indoubt transaction.
-     * Cloud002, a peer server as it belongs to the same recovery group is started and recovery the
-     * transaction that belongs to Cloud001.
-     *
-     * @throws Exception
-     */
-    @Test
-    @AllowedFFDC(value = { "com.ibm.ws.recoverylog.spi.RecoveryFailedException" })
-    public void testDBRecoveryTakeover() throws Exception {
-        final String method = "testDBRecoveryTakeover";
-        StringBuilder sb = null;
-        String id = "001";
-        // Start Server1
-        server1.startServer();
-
-        try {
-            // We expect this to fail since it is gonna crash the server
-            sb = runTestWithResponse(server1, SERVLET_NAME, "setupRec" + id);
-        } catch (Throwable e) {
-        }
-        Log.info(this.getClass(), method, "setupRec" + id + " returned: " + sb);
-
-        server1.waitForStringInLog("Dump State:");
-
-        // Now start server2
-        ProgramOutput po = server2.startServerAndValidate(false, true, true);
-        if (po.getReturnCode() != 0) {
-            Log.info(this.getClass(), method, po.getCommand() + " returned " + po.getReturnCode());
-            Log.info(this.getClass(), method, "Stdout: " + po.getStdout());
-            Log.info(this.getClass(), method, "Stderr: " + po.getStderr());
-            Exception ex = new Exception("Could not restart the server");
-            Log.error(this.getClass(), "recoveryTest", ex);
-            throw ex;
-        }
-
-        // Server appears to have started ok. Check for key string to see whether peer recovery has succeeded
-        server2.waitForStringInTrace("Performed recovery for cloud001");
-        // "CWWKE0701E" error message is allowed
-        server2.stopServer("CWWKE0701E");
-
-        // Lastly, clean up XA resource files
-        server1.deleteFileFromLibertyInstallRoot("/usr/shared/" + LastingXAResourceImpl.STATE_FILE_ROOT);
-    }
-
-    /**
-     * The purpose of this test is to verify correct behaviour when peer servers compete for a log.
-     *
-     * The Cloud001 server is started and a servlet invoked. The servlet modifies the owner of the server's
-     * lease recored in the lease table. This simulates the situation where a peer server has acquired the
-     * ownership of the lease and is recovering Cloud001's logs. Finally the servlet halts the server leaving
-     * an indoubt transaction.
-     *
-     * Cloud001 is restarted but should fail to acquire the lease to its recovery logs as it is no longer the owner.
-     *
-     * @throws Exception
-     */
-    @Test
-    @ExpectedFFDC(value = { "com.ibm.ws.recoverylog.spi.RecoveryFailedException" })
-    @AllowedFFDC(value = { "javax.transaction.xa.XAException", "com.ibm.tx.jta.XAResourceNotAvailableException", "com.ibm.ws.recoverylog.spi.RecoveryFailedException",
-                           "java.lang.IllegalStateException" })
-    // defect 227411, if cloud002 starts slowly, then access to cloud001's indoubt tx
-    // XAResources may need to be retried (tx recovery is, in such cases, working as designed.
-    public void testDBRecoveryCompeteForLog() throws Exception {
-        final String method = "testDBRecoveryCompeteForLog";
-        StringBuilder sb = null;
-        String id = "001";
-
-        // Start Server1
-        longLeaseLengthServer1.startServer();
-
-        try {
-            sb = runTestWithResponse(longLeaseLengthServer1, SERVLET_NAME, "modifyLeaseOwner");
-
-            // We expect this to fail since it is gonna crash the server
-            sb = runTestWithResponse(longLeaseLengthServer1, SERVLET_NAME, "setupRec" + id);
-        } catch (Throwable e) {
-        }
-        Log.info(this.getClass(), method, "setupRec" + id + " returned: " + sb);
-
-        longLeaseLengthServer1.waitForStringInLog("Dump State:");
-        longLeaseLengthServer1.postStopServerArchive(); // must explicitly collect since crashed server
-        // Pull in a new server.xml file that ensures that we have a long (5 minute) timeout
-        // for the lease, otherwise we may decide that we CAN delete and renew our own lease.
-
-        // Now re-start cloud1 but we fully expect this to fail
-        try {
-            longLeaseLengthServer1.startServerExpectFailure("recovery-dblog-fail.log", false, true);
-        } catch (Exception ex) {
-            // Tolerate an exception here, as recovery is asynch and the "successful start" message
-            // may have been produced by the main thread before the recovery thread had completed
-            Log.info(this.getClass(), method, "startServerExpectFailure threw exc: " + ex);
-        }
-
-        // Server appears to have failed as expected. Check for log failure string
-        if (longLeaseLengthServer1.waitForStringInLog("RECOVERY_LOG_FAILED") == null) {
-            Exception ex = new Exception("Recovery logs should have failed");
-            Log.error(this.getClass(), "recoveryTestCompeteForLock", ex);
-            throw ex;
-        }
-        longLeaseLengthServer1.postStopServerArchive(); // must explicitly collect since server start failed
-        // defect 210055: Now start cloud2 so that we can tidy up the environment, otherwise cloud1
-        // is unstartable because its lease is owned by cloud2.
-        ProgramOutput po = server2.startServerAndValidate(false, true, true);
-        if (po.getReturnCode() != 0) {
-            Log.info(this.getClass(), method, po.getCommand() + " returned " + po.getReturnCode());
-            Log.info(this.getClass(), method, "Stdout: " + po.getStdout());
-            Log.info(this.getClass(), method, "Stderr: " + po.getStderr());
-            Exception ex = new Exception("Could not restart the server");
-            Log.error(this.getClass(), "recoveryTest", ex);
-            throw ex;
-        }
-
-        // Server appears to have started ok. Check for 2 key strings to see whether peer recovery has succeeded
-        server2.waitForStringInTrace("Performed recovery for cloud001");
-        // "CWWKE0701E" error message is allowed
-        server2.stopServer("CWWKE0701E");
-
-        // Lastly, clean up XA resource files
-        server1.deleteFileFromLibertyInstallRoot("/usr/shared/" + LastingXAResourceImpl.STATE_FILE_ROOT);
-    }
-}
