--- dev/com.ibm.ws.transaction.cloud_fat.simpleDB/publish/servers/CLOUD002/bootstrap.properties
@@ -1,5 +1,0 @@
-bootstrap.include=../testports.properties
-com.ibm.ws.logging.trace.specification=com.ibm.tx.jta.util.TxTMHelper=all=enabled
-ds.loglevel=DEBUG
-com.ibm.tx.auditRecovery=true
-osgi.console=5471
