--- dev/com.ibm.ws.transaction.cloud_fat.simpleDB/transaction_fat.bnd
@@ -1,15 +1,0 @@
-#*******************************************************************************
-# Copyright (c) 2018,2020 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License v1.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-v10.html
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-#*******************************************************************************
--include= ~../cnf/resources/bnd/bundle.props
-bVersion=1.0
-
-# Define the bundle for this FAT
-Bundle-SymbolicName: ${basename;${basedir}}
